Bind httpd to all interfaces
This change is needed we stopped adding "127.0.0.1 {{ internal_host }}"
to /etc/hosts.  Since haproxy uses the host name as the route to httpd,
this change to /etc/hosts means the host name no longer leads to
localhost, which means haproxy can't talk to httpd anymore.  And I want
haproxy to use the host name so it can verify TLS. Hence the need to
bind httpd to all interfaces.