Take the LDAP URL as-is from the provided facts
The deployer must know the correct LDAP URL syntax for this site;
specifically, whether it should be a list or a string.  The installer
will now use the ldap_url from the facts, and will not attempt to decide
whether to stringify it, convert a list to a comma separated string, or
make it a list, or any other transformation.